backend/requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
django-notifications-hq==1.8.3
setuptools==69.2.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.31.0
pdfrw==0.4
python-dateutil==2.8.2
django-apscheduler==0.6.2
pytz==2023.3.post1

